Init onboarding flow
Enables logging in via LinkedIn, and being directed to either the
home page or the new profile form if the user has not onboarded yet.
When users sign up or first log in with LinkedIn, they will need to
create a profile.

In a future commit, we can build on this with more attributes and a form
that saves the profile.